Check for data.tree >= v1.1.1 for data generation
This is the development version from the pull request gluc/data.tree#174
As we used this version to generate the culture trees in srppp_list,
there is no warning during R CMD check any more, so we can error on
warning again in the github workflow. Also, installation does not give
warnings any more. Closes #2.