Function for finding unauthorized products
Function to check if the product name indicates via a comment that the
use of the product is not authorized in the year of the publication of
the srppp. This mainly finds products in 2011 that had no active
authorization but were still included.

The use of the function is demonstrated in the vignette about trends of
active substances over the years.

Closes #3